feat: 구간 distance 필드를 Long 으로

+++ src/main/java/kuit/subway/dto/request/ModifyLineRequest.java
@@ -13,7 +13,7 @@ public class ModifyLineRequest {
     @NotNull(message = "지하철 노선의 색상은 Null 일 수 없습니다.")
     private String color;
     @NotNull(message = "지하철 노선의 거리는 Null 일 수 없습니다.")
-    private int distance;
+    private Long distance;
     @Size(min = 3, max = 20)
     private String name;
     @NotNull(message = "지하철 노선의 하행 종점역은 정의되어야 합니다.")
